+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### NEXT
 
-- Simulcast and SVC: Limit temporal layer to the preferred one ([PR #1892](https://github.com/versatica/mediasoup/pull/1892)).
+- **Breaking change:** Simulcast and SVC: Limit temporal layer to the preferred one ([PR #1892](https://github.com/versatica/mediasoup/pull/1892)).
 
 ### 3.25.0