+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### NEXT
 
+### 3.26.0
+
 - **Breaking change:** Simulcast and SVC: Limit temporal layer to the preferred one ([PR #1892](https://github.com/versatica/mediasoup/pull/1892)).
 
 ### 3.25.0